requirements.txt
NBT
mariadb
python-dotenv